+++ packages/utils/src/converters/LlmSchemaConverter.ts
@@ -11,7 +11,7 @@ import { NamingConvention } from "../utils/NamingConvention";
 import { JsonDescriptor } from "../utils/internal/JsonDescriptor";
 import { LlmTypeChecker } from "../validators/LlmTypeChecker";
 import { OpenApiTypeChecker } from "../validators/OpenApiTypeChecker";
-import { OpenApiValidator } from "../validators/OpenApiValidator";
+import { OpenApiValidator } from "../validators/internal/OpenApiValidator";
 import { LlmDescriptionInverter } from "./internal/LlmDescriptionInverter";
 import { LlmParametersFinder } from "./internal/LlmParametersComposer";
 import { OpenApiConstraintShifter } from "./internal/OpenApiConstraintShifter";
